Simplify handlers
